feat: added splashscreen documentation: added comments and TODOs changes: moved db to outside app folder
