+++ project1/app/(tab)/[details].tsx
@@ -1,0 +1,1 @@
+// basically dynamically show a deailts page of the city based on what citiy we press on the cards on press should show this page too
