+++ project1/app/(tab)/favorites.tsx
@@ -7,5 +7,6 @@ export default function FavoriteScreen(){
 
     return (
         <View><Text>Favorites Page</Text></View>
+        // The goal is to have favourites show up here maybe use the weathercard component
     );
 }
